requirements.txt
streamlit==1.46.0
pillow==11.2.1
pandas==2.2.3
numpy==2.2.5
torch==2.7.1
torchvision==0.22.1
timm==1.0.16
faiss-cpu==1.11.0
openai==1.97.0
